separate dummy actions from VRInput handler; Radial menu should probably use this too.
