+++ GTFO_VR/Core/VR_Input/SteamVR_InputHandler.cs
@@ -48,16 +48,6 @@ namespace GTFO_VR.Core.VR_Input
 
         private static Dictionary<InputAction, SteamVR_Action_Boolean> boolActions;
 
-        private static DummyAction m_terminalDeleteAction= new DummyAction(InputAction.TerminalDel, true);
-        private static DummyAction m_terminalExitAction = new DummyAction(InputAction.TerminalExit, true);
-
-        private static DummyAction m_terminalLeftAction = new DummyAction(InputAction.TerminalLeft, true);
-        private static DummyAction m_terminalUpAction = new DummyAction(InputAction.TerminalUp, true);
-        private static DummyAction m_terminalRightAction = new DummyAction(InputAction.TerminalRight, true);
-        private static DummyAction m_terminalDownAction = new DummyAction(InputAction.TerminalDown, true);
-
-        private static Dictionary<InputAction, DummyAction> dummyBoolActions;
-
         public static void Setup()
         {
             if(Initialized)
@@ -105,25 +95,12 @@ namespace GTFO_VR.Core.VR_Input
             hapticAction.Execute(0.0f, duration, frequency, amplitude, controller);
         }
 
-        public static void triggerDummyAction(InputAction action )
-        {
-            DummyAction dummyAction = GetDummyBoolActionMapping(action);
-            if (dummyAction != null )
-            {
-                dummyAction.requestPress();
-            }
-        }
-
         public static bool GetActionDown(InputAction action)
         {
             if (!Initialized)
             {
                 return false;
             }
-
-            DummyAction dummyAction = GetDummyBoolActionMapping(action);
-            if (dummyAction != null && dummyAction.getDown())
-                return true;
 
             SteamVR_Action_Boolean boolActionMapping = GetBoolActionMapping(action);
             if (IsIRLCrouchValid(action))
@@ -149,10 +126,6 @@ namespace GTFO_VR.Core.VR_Input
                 return false;
             }
 
-            //DummyAction dummyAction = GetDummyBoolActionMapping(action);
-            //if (dummyAction != null && dummyAction.getDown())
-            //    return true;
-
             SteamVR_Action_Boolean boolActionMapping = SteamVR_InputHandler.GetBoolActionMapping(action);
             if (IsIRLCrouchValid(action))
             {
@@ -167,10 +140,6 @@ namespace GTFO_VR.Core.VR_Input
             {
                 return false;
             }
-
-            DummyAction dummyAction = GetDummyBoolActionMapping(action);
-            if (dummyAction != null && dummyAction.getUp())
-                return true;
 
             SteamVR_Action_Boolean boolActionMapping = GetBoolActionMapping(action);
             if (IsIRLCrouchValid(action))
@@ -240,17 +209,6 @@ namespace GTFO_VR.Core.VR_Input
                 { InputAction.Flashlight, m_flashlightAction },
             };
 
-            dummyBoolActions = new Dictionary<InputAction, DummyAction>
-            {
-                { InputAction.TerminalDel, m_terminalDeleteAction },
-                { InputAction.TerminalExit, m_terminalExitAction },
-
-                { InputAction.TerminalLeft, m_terminalLeftAction },
-                { InputAction.TerminalUp, m_terminalUpAction },
-                { InputAction.TerminalRight, m_terminalRightAction },
-                { InputAction.TerminalDown, m_terminalDownAction }
-            };
-
         }
 
         private static SteamVR_Action_Boolean GetBoolActionMapping(InputAction action)
@@ -260,15 +218,6 @@ namespace GTFO_VR.Core.VR_Input
                 return boolActions[action];
             }
 
-            return null;
-        }
-
-        private static DummyAction GetDummyBoolActionMapping(InputAction action)
-        {
-            if (dummyBoolActions.ContainsKey(action))
-            {
-                return dummyBoolActions[action];
-            }
             return null;
         }
