+++ GTFO_VR/Injections/Input/InjectInput.cs
@@ -25,7 +25,7 @@ namespace GTFO_VR.Injections.Input
     {
         private static void Postfix(InputAction action, ref bool __result)
         {
-            __result = __result || SteamVR_InputHandler.GetActionUp(action);
+            __result = __result || SteamVR_InputHandler.GetActionUp(action) || Dummy_InputHandler.GetActionUp(action); ;
         }
     }
 
@@ -34,7 +34,7 @@ namespace GTFO_VR.Injections.Input
     {
         private static void Postfix(InputAction action, ref bool __result)
         {
-            __result = __result || SteamVR_InputHandler.GetActionDown(action) || WeaponRadialMenu.GetSpecialActionMappingDown(action);
+            __result = __result || SteamVR_InputHandler.GetActionDown(action) || Dummy_InputHandler.GetActionDown(action) || WeaponRadialMenu.GetSpecialActionMappingDown(action);
         }
     }
 
@@ -43,7 +43,7 @@ namespace GTFO_VR.Injections.Input
     {
         private static void Postfix(InputAction action, ref bool __result)
         {
-            __result = __result || SteamVR_InputHandler.GetAction(action) || WeaponRadialMenu.GetSpecialActionMappingDown(action);
+            __result = __result || SteamVR_InputHandler.GetAction(action) || Dummy_InputHandler.GetActionDown(action) ||  WeaponRadialMenu.GetSpecialActionMappingDown(action);
         }
     }
 }
